pyproject.toml
[project]
name = "fruit-classifier-backend"
version = "0.1.0"
description = "A FastAPI backend for fruit classification"
authors = [
    {name = "Newt Louis", email = "l.ngoc.17395@gmail.com" },
    {name = "Jun", email = "jun.coder95@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "kaggle (>=1.7.4.5,<2.0.0.0)",
    "kagglehub (>=0.3.12,<0.4.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "torch (>=2.7.1,<3.0.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.7.1,<2.0.0)"
]
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
packages = [{ include = "src" }]

[tool.poetry.group.dev.dependencies]
kaggle = "^1.7.4.5"

